Add stub source as a development/test plugin
