+++ cobra_commander/spec/fixtures/app_helper.rb
@@ -3,16 +3,6 @@
 require "yaml"
 
 module AppHelper
-  class StubSource < CobraCommander::Source[:stub]
-    def packages
-      @packages ||= YAML.load_file("#{path}.yml").map do |name, dependencies|
-        CobraCommander::Package.new(self, name: name,
-                                          path: path.join(name),
-                                          dependencies: dependencies || [])
-      end
-    end
-  end
-
   def stub_umbrella(path = nil)
     CobraCommander.umbrella(fixture_file_path(path), stub: true)
   end
