+++ cobra_commander/spec/fixtures/cobra_commander-stub.gemspec
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+Gem::Specification.new do |spec|
+  spec.name = "cobra_commander-stub"
+  spec.version = "0.1.0"
+  spec.summary = "cobra_commander-stub"
+  spec.authors = ["Carlos Palhares"]
+
+  spec.files = ["stub_source.rb"]
+  spec.require_paths = ["./"]
+end
